Adds named enums to openapi document
