+++ src/openapi/schemas.js
@@ -1,0 +1,7 @@
+import allGenders from "../data/genders.js";
+import allProvinces from "../data/provinces.js";
+
+export default {
+  Gender: { enum: allGenders },
+  Province: { enum: allProvinces },
+};
